Add mobileOnly section for label and mode to the Dynamic Preferences example.
When SmartApps use dynamic pages, label and mode properties are not
added automatically to the app's preferences. This change adds the
fields to the Dynamic Preferences example. Initially I was going to
call this a bug, but seeing how the official apps do this, I think
at the very least it needs to be documented.